requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
sentence-transformers==2.3.1
prometheus-client==0.19.0
python-dotenv==1.0.1
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
langchain==0.1.6
langchain-community==0.0.19
torch==2.2.0
faiss-cpu==1.7.4
pydantic==2.6.0
openai==1.47.0
scikit-learn==1.3.2